Review changes
A lot of changes are done after installing checkstyle plugin. I worked
on my packages. Removed test from PositionCalculatorTest after changing
a way that method parseInt works. Now instead throwing excpetion returns
-1 which is number that handles board class.
Added a lot of java docs.
Still there is bug with horizontal win checker. Fixing it is my next
goal.
