+++ pom.xml
@@ -18,10 +18,14 @@
         <maven.compiler.target>${project.java.version}</maven.compiler.target>
         <maven.compiler.encoding>${project.build.sourceEncoding}</maven.compiler.encoding>
         <maven.resources.encoding>${project.build.sourceEncoding}</maven.resources.encoding>
-        <version.plugin.maven.jacoco>0.8.5</version.plugin.maven.jacoco>
         <version.maven>3.6.3</version.maven>
         <version.plugin.maven.enforcer>3.0.0-M2</version.plugin.maven.enforcer>
-
+        <version.plugin.maven.jar>3.1.0</version.plugin.maven.jar>
+        <version.plugin.maven.compiler>3.8.0</version.plugin.maven.compiler>
+        <version.plugin.maven.resources>3.1.0</version.plugin.maven.resources>
+        <version.plugin.maven.jacoco>0.8.5</version.plugin.maven.jacoco>
+        <version.dependency.testng>7.1.0</version.dependency.testng>
+        <version.plugin.maven.checkstyle>3.1.0</version.plugin.maven.checkstyle>
         <boardSize>15</boardSize>
     </properties>
 
@@ -29,7 +33,7 @@
         <dependency>
             <groupId>org.testng</groupId>
             <artifactId>testng</artifactId>
-            <version>7.0.0</version>
+            <version>${version.dependency.testng}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -132,7 +136,7 @@
             <!-- Declare jar plugin -->
             <plugin>
                 <artifactId>maven-jar-plugin</artifactId>
-                <version>3.0.2</version>
+                <version>${version.plugin.maven.jar}</version>
                 <configuration>
                     <archive>
                         <manifest>
@@ -172,6 +176,27 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-pmd-plugin</artifactId>
                 <version>3.12.0</version>
+            </plugin>
+            <!-- Checkstyle -->
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-checkstyle-plugin</artifactId>
+                <version>${version.plugin.maven.checkstyle}</version>
+                <configuration>
+                    <configLocation>google_checks.xml</configLocation>
+                    <encoding>UTF-8</encoding>
+                    <consoleOutput>true</consoleOutput>
+                    <failsOnError>true</failsOnError>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>validate</id>
+                        <phase>validate</phase>
+                        <goals>
+                            <goal>check</goal>
+                        </goals>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
