+++ runAutoTest.sh
@@ -1,2 +1,4 @@
 ./auto.sh >> ./log.txt
+cat ./log.txt
+echo "______________________________"
 echo "Test is stored in file log.txt"
