+++ src/main/java/com/bieganski/ox/GameListener.java
@@ -1,5 +1,16 @@
 package com.bieganski.ox;
 
+/**
+ * Interface to notify about game status changes.
+ */
 public interface GameListener {
-    void onWin();
+  /**
+   * Notify about pass win conditions.
+   */
+  void onWin();
+
+  /**
+   * Notify about draw
+   */
+  void onDraw();
 }
