--- src/main/java/com/bieganski/ox/Player.java
@@ -1,11 +1,0 @@
-package com.bieganski.ox;
-
-public class Player {
-    private String name;
-    private int score;
-    private Symbol symbol;
-
-    public Player(String name) {
-        this.name = name;
-    }
-}
