--- src/main/java/com/bieganski/ox/PositionValidator.java
@@ -1,14 +1,0 @@
-package com.bieganski.ox;
-
-import java.util.regex.Pattern;
-
-class PositionValidator implements Validator {
-
-    @Override
-    public boolean checkString(String textToValidate) {
-        return Pattern
-                .compile("^[0-9]+ [0-9]+$")
-                .matcher(textToValidate)
-                .find();
-    }
-}
