--- src/main/java/com/bieganski/ox/Symbol.java
@@ -1,5 +1,0 @@
-package com.bieganski.ox;
-
-enum Symbol {
-    X, O
-}
