--- src/main/java/com/bieganski/ox/Validator.java
@@ -1,5 +1,0 @@
-package com.bieganski.ox;
-
-interface Validator {
-    boolean checkString(String textToValidate);
-}
