+++ src/main/java/com/bieganski/ox/model/Symbol.java
@@ -1,0 +1,8 @@
+package com.bieganski.ox.model;
+
+/**
+ * Symbols that player can place on board.
+ */
+public enum Symbol {
+  X, O
+}
