--- src/main/java/com/bieganski/ox/painters/BoardPainter.java
@@ -1,9 +1,0 @@
-package com.bieganski.ox.painters;
-
-import com.bieganski.ox.Field;
-
-import java.util.TreeSet;
-
-public interface BoardPainter {
-    String paintBoard(TreeSet<Field> fieldsWithValues, int boardSide);
-}
