+++ src/main/java/com/bieganski/ox/ui/BoardPainter.java
@@ -1,0 +1,9 @@
+package com.bieganski.ox.ui;
+
+import com.bieganski.ox.model.Field;
+
+import java.util.TreeSet;
+
+interface BoardPainter {
+  String paintBoard(TreeSet<Field> fieldsWithValues, int boardSide);
+}
