+++ src/main/java/com/bieganski/ox/ui/PositionValidator.java
@@ -1,0 +1,14 @@
+package com.bieganski.ox.ui;
+
+import java.util.regex.Pattern;
+
+class PositionValidator implements Validator {
+
+  @Override
+  public boolean checkString(String textToValidate) {
+    return Pattern
+        .compile("^[0-9]+ [0-9]+$")
+        .matcher(textToValidate)
+        .find();
+  }
+}
