--- src/main/java/com/bieganski/ox/user_interface/UserInterface.java
@@ -1,9 +1,0 @@
-package com.bieganski.ox.user_interface;
-
-import com.bieganski.ox.BoardListener;
-
-public interface UserInterface extends BoardListener {
-    void println(Object object);
-
-    String askForInput();
-}
