--- src/main/java/com/bieganski/ox/win_checkers/WinChecker.java
@@ -1,14 +1,0 @@
-package com.bieganski.ox.win_checkers;
-
-import com.bieganski.ox.BoardListener;
-import com.bieganski.ox.Field;
-
-import java.util.TreeSet;
-
-/**
- * Interface for win-check classes
- * @author Kamil Biegański
- */
-interface WinChecker {
-    boolean checkWin(TreeSet<Field> fieldsWithValue, Field addedField, int size);
-}
