+++ src/test/java/com/bieganski/ox/ui/ConsoleBoardPainterTest.java
@@ -1,6 +1,7 @@
-package com.bieganski.ox;
+package com.bieganski.ox.ui;
 
-import com.bieganski.ox.painters.ConsoleBoardPainter;
+import com.bieganski.ox.model.Field;
+import com.bieganski.ox.model.Symbol;
 import org.testng.annotations.Test;
 
 import java.util.TreeSet;
